refactor to one class per module
