+++ sparrow_datums/boxes/__init__.py
@@ -1,4 +1,7 @@
+from .augmented_boxes import AugmentedBoxes
 from .boxes import Boxes
-from .frame_boxes import FrameAugmentedBoxes, FrameBoxes
+from .frame_augmented_boxes import FrameAugmentedBoxes
+from .frame_boxes import FrameBoxes
 from .iou import pairwise_iou
-from .single_box import SingleAugmentedBox, SingleBox
+from .single_augmented_box import SingleAugmentedBox
+from .single_box import SingleBox
