+++ sparrow_datums/boxes/single_box_test.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from ..types import PType
-from .single_box import SingleAugmentedBox, SingleBox
+from .single_box import SingleBox
 
 
 def test_single_box_conversion_creates_single_box():
@@ -17,17 +17,3 @@ def test_single_box_with_bad_shape_throws_value_error():
         SingleBox(np.ones((2, 4)))
     with pytest.raises(ValueError):
         SingleBox(np.ones(3))
-
-
-def test_single_augmented_box_conversion_creates_single_augmented_box():
-    box_a = SingleAugmentedBox(np.ones(6), PType.absolute_tlwh)
-    box_b = box_a.to_tlbr()
-    assert box_b.is_tlbr
-    assert isinstance(box_b, SingleAugmentedBox)
-
-
-def test_single_augmented_box_with_bad_shape_throws_value_error():
-    with pytest.raises(ValueError):
-        SingleAugmentedBox(np.ones((2, 6)))
-    with pytest.raises(ValueError):
-        SingleAugmentedBox(np.ones(4))
